+++ app/models/auth.rb
@@ -1,0 +1,4 @@
+class Auth
+  include Lotus::Entity
+  self.attributes = :user_id, :provider, :external_id, :email, :token, :secret
+end