+++ app/models/user.rb
@@ -1,0 +1,4 @@
+class User
+  include Lotus::Entity
+  self.attributes = :email
+end