+++ app/views/auth/failure.rb
@@ -1,4 +1,4 @@
-module GoSpotify::Views::Session
+module GoSpotify::Views::Auth
   class Failure
     include GoSpotify::View
   end