+++ application.rb
@@ -1,3 +1,5 @@
+APP_ENV = ENV["RACK_ENV"] || "development"
+
 require "lotus"
 require "lotus/action/session"
 require "lotus-dynamodb"
@@ -6,10 +8,22 @@ require "omniauth-spotify"
 require "omniauth-soundcloud"
 require "omniauth-vkontakte"
 
-AWS.config(
-  access_key_id: ENV["AWS_KEY"],
-  secret_access_key: ENV["AWS_SECRET"],
-)
+# :nocov:
+if APP_ENV == "production"
+  AWS.config(
+    access_key_id: ENV["AWS_KEY"],
+    secret_access_key: ENV["AWS_SECRET"],
+  )
+else
+  AWS.config(
+    use_ssl: false,
+    dynamo_db_endpoint: "localhost",
+    dynamo_db_port: 4567,
+    access_key_id: "",
+    secret_access_key: "",
+  )
+end
+# :nocov:
 
 module GoSpotify
   module CommonAction
@@ -25,7 +39,7 @@ module GoSpotify
 
         expose :current_user, :user_signed_in
         def current_user
-          @current_user ||= "lol" if session[:user_id]
+          @current_user ||= UserRepository.find(session[:user_id]) if session[:user_id]
         end
 
         def user_signed_in
@@ -43,11 +57,12 @@ module GoSpotify
     configure do
       layout :application
       load_paths << "app"
-      routes  "config/routes"
-      mapping "config/mapping"
+      routes "config/routes"
     end
   end
 end
 
 APP = GoSpotify::Application.new
-LOGGER = Logger.new("log/development.log")
+LOGGER = Logger.new("log/#{APP_ENV}.log")
+
+require_relative "config/mapping"