+++ config/routes.rb
@@ -1,4 +1,4 @@
-get "/",                         to: "home#index", as: :root
-get "/auth/:provider/callback",  to: "session#callback"
-get "/auth/failure",             to: "session#failure"
-get "/auth/signout",             to: "session#signout"
+get "/",                         to: "home#index",    as: :root
+get "/auth/:provider/callback",  to: "auth#callback", as: :auth_callback
+get "/auth/failure",             to: "auth#failure",  as: :auth_failure
+get "/auth/signout",             to: "auth#signout",  as: :auth_signout