+++ test/test_helper.rb
@@ -1,3 +1,5 @@
+ENV["RACK_ENV"] ||= "test"
+
 require "rubygems"
 require "bundler/setup"
 
@@ -19,8 +21,15 @@ end
 require "minitest/autorun"
 require "minitest/capybara"
 require "minitest/features"
+require "rack_session_access"
+require "rack_session_access/capybara"
 require "tilt/erb"
 require_relative "../application"
+
+def before_each
+  AuthRepository.clear
+  UserRepository.clear
+end
 
 Capybara.app = Rack::Builder.parse_file(File.expand_path("../../config.ru", __FILE__)).first
 Capybara.javascript_driver = :webkit
@@ -28,18 +37,20 @@ Capybara.javascript_driver = :webkit
 OmniAuth.config.logger.level = Logger::FATAL
 OmniAuth.config.test_mode = true
 OmniAuth.config.add_mock(:spotify, {
-  uid: "12345",
+  uid: "john",
   info: {
-    name: "John Smith",
     email: "john@smith.com",
-  }
+  },
+  credentials: {
+    token: "XXX",
+    refresh_token: "XXX",
+  },
 })
-
-AWS.config(
-  use_ssl: false,
-  dynamo_db_endpoint: "localhost",
-  dynamo_db_port: 4567,
-  access_key_id: "",
-  secret_access_key: "",
-)
-Net::HTTP.new("localhost", AWS.config.dynamo_db_port).delete("/")
+OmniAuth.config.add_mock(:soundcloud, {
+  uid: 12345,
+  info: {
+  },
+  credentials: {
+    token: "XXX",
+  },
+})