+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 2014-07-02
 ----------
 
-* Initial version: home page, Spotify auth, signout
+* Implement auth
+* Initial version: home page, dummy auth, signout